initial prototype of lrpar codegen
This had to introduce a call to `parser_builder.from_ast.clone()`
to work.
